🚚 Reorganize specs
- Unit tests separate from system tests
- All test supporting ruby and artifacts organized in support folders
- Common test cases reorganized to their own files with modules renamed
- Supporting Ruby decomposed into individual classes and files
