+++ .rspec
@@ -1,0 +1,2 @@
+-I spec/support
+-I spec/support/system
