+++ spec/support/system/gem_dir_layout.rb
@@ -1,0 +1,20 @@
+# =========================================================================
+#   Ceedling - Test-Centered Build System for C
+#   ThrowTheSwitch.org
+#   Copyright (c) 2010-25 Mike Karlesky, Mark VanderVoord, & Greg Williams
+#   SPDX-License-Identifier: MIT
+# =========================================================================
+
+class GemDirLayout
+  attr_reader :gem_dir_base_name
+
+  def initialize(install_dir)
+    @gem_dir_base_name = "gems"
+    @d = File.join install_dir, @gem_dir_base_name
+    FileUtils.mkdir_p @d
+  end
+
+  def install_dir; convert_slashes(@d)  end
+  def bin;         File.join(@d, 'bin') end
+  def lib;         File.join(@d, 'lib') end
+end
