+++ spec/units/bin/merginator_spec.rb
