+++ spec/units/c_extractor/c_extractor_code_text_spec.rb
