+++ spec/units/c_extractor/c_extractor_declarations_spec.rb
