+++ spec/units/c_extractor/c_extractor_functions_spec.rb
