+++ spec/units/c_extractor/c_extractor_integration_spec.rb
