+++ spec/units/c_extractor/c_extractor_spec.rb
