+++ spec/units/config_walkinator_spec.rb
