+++ spec/units/configurator_builder_spec.rb
