+++ spec/units/configurator_helper_spec.rb
