+++ spec/units/configurator_spec.rb
