+++ spec/units/generators/generator_partials_spec.rb
