+++ spec/units/generators/generator_test_results_sanity_checker_spec.rb
