+++ spec/units/generators/generator_test_results_spec.rb
