+++ spec/units/partials/partializer_config_spec.rb
