+++ spec/units/preprocess/c_comment_scanner_spec.rb
