+++ spec/units/preprocess/preprocessinator_comment_stripper_spec.rb
