+++ spec/units/reportinator_spec.rb
