+++ spec/units/test_context_extractor_spec.rb
